requirements.txt
Flask
Werkzeug
requests
pandas
numpy
yfinance
scipy
reportlab
Flask-Limiter==3.5.0
cryptography==41.0.7
pyjwt==2.8.0
python-dotenv==1.0.0
flask-talisman==1.1.0
flask-sslify==0.1.5 